Update sasquatch apps to use eventfilter service
