+++ TestApps/Sasquatch/Sasquatch/Sasquatch-Bridging-Header.h
@@ -4,4 +4,4 @@
 
 #import "Constants.h"
 #import "CrashLib.h"
-#import "MSLogFilter.h"
+#import "MSEventFilter.h"
