+++ TestApps/SasquatchMac/SasquatchMac/AppCenterDelegate.swift
@@ -58,10 +58,7 @@ import AppCenter
   func lastCrashReportDeviceCarrierCountry() -> String?
   func lastCrashReportDeviceAppNamespace() -> String?
 
-  // MSLogFilter section.
-  func isLogFilterEnabled() -> Bool
-  func setLogFilterEnabled(_ isEnabled: Bool)
-  func filterLogType(_ logType: String)
-  func unfilterLogType(_ logType: String)
-  func isFilteringLogType(_ logType: String) -> Bool
+  // MSEventFilter section.
+  func isEventFilterEnabled() -> Bool
+  func setEventFilterEnabled(_ isEnabled: Bool)
 }
