+++ TestApps/SasquatchMac/SasquatchMac/EventFilter/InternalHeaders/MSChannelConfiguration.h
