+++ TestApps/SasquatchMac/SasquatchMac/EventFilter/InternalHeaders/MSLogManager.h
