+++ TestApps/SasquatchMac/SasquatchMac/EventFilter/InternalHeaders/MSLogManagerDelegate.h
