+++ TestApps/SasquatchMac/SasquatchMac/EventFilter/MSEventFilter.h
@@ -1,0 +1,12 @@
+#import <AppCenter/MSServiceAbstract.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+/**
+ * Event filtering service.
+ */
+@interface MSEventFilter : MSServiceAbstract
+@end
+
+NS_ASSUME_NONNULL_END
+
