+++ TestApps/SasquatchMac/SasquatchMac/SasquatchMac-Bridging-Header.h
@@ -3,4 +3,4 @@
 //
 
 #import "CrashLib.h"
-#import "MSLogFilter.h"
+#import "MSEventFilter.h"
