+++ TestApps/SasquatchMac/SasquatchMacObjC/AppCenterDelegateObjC.m
@@ -1,5 +1,5 @@
 #import "AppCenterDelegateObjC.h"
-#import "MSLogFilter.h"
+#import "MSEventFilter.h"
 
 @import AppCenter;
 @import AppCenterAnalytics;
@@ -150,21 +150,12 @@
   return [[[MSCrashes lastSessionCrashReport] device] carrierCountry];
 }
 
-#pragma mark - MSLogFilter section.
-- (BOOL) isLogFilterEnabled {
-  return [MSLogFilter isEnabled];
+#pragma mark - MSEventFilter section.
+- (BOOL) isEventFilterEnabled {
+  return [MSEventFilter isEnabled];
 }
-- (void) setLogFilterEnabled:(BOOL)isEnabled {
-  [MSLogFilter setEnabled:isEnabled];
-}
-- (void) filterLogType:(NSString*)logType {
-  [MSLogFilter filterLogType:logType];
-}
-- (void) unfilterLogType:(NSString*)logType {
-  [MSLogFilter unfilterLogType:logType];
-}
-- (BOOL) isFilteringLogType:(NSString*)logType {
-  return [MSLogFilter isFilteringLogType:logType];
+- (void) setEventFilterEnabled:(BOOL)isEnabled {
+  [MSEventFilter setEnabled:isEnabled];
 }
 
 @end
