+++ TestApps/SasquatchMac/SasquatchMacSwift/AppCenterDelegateSwift.swift
@@ -147,20 +147,11 @@ class AppCenterDelegateSwift : AppCenterDelegate {
     return MSCrashes.lastSessionCrashReport()?.device.appNamespace
   }
 
-  //MARK: MSLogFilter section.
-  func isLogFilterEnabled() -> Bool{
-    return MSLogFilter.isEnabled();
+  //MARK: MSEventFilter section.
+  func isEventFilterEnabled() -> Bool{
+    return MSEventFilter.isEnabled();
   }
-  func setLogFilterEnabled(_ isEnabled: Bool){
-    MSLogFilter.setEnabled(isEnabled);
-  }
-  func filterLogType(_ logType: String){
-    MSLogFilter.filterLogType(logType);
-  }
-  func unfilterLogType(_ logType: String){
-    MSLogFilter.unfilterLogType(logType);
-  }
-  func isFilteringLogType(_ logType: String) -> Bool {
-    return MSLogFilter.isFilteringLogType(logType)
+  func setEventFilterEnabled(_ isEnabled: Bool){
+    MSEventFilter.setEnabled(isEnabled);
   }
 }
