+++ TestApps/SasquatchTV/SasquatchTV/SasquatchTV-Bridging-Header.h
@@ -4,4 +4,4 @@
 
 #import "CrashLib.h"
 #import "MSAlertController.h"
-#import "MSLogFilter.h"
+#import "MSEventFilter.h"
