+++ TestApps/SasquatchTV/SasquatchTVObjC/AppCenterDelegateObjC.m
@@ -1,6 +1,6 @@
 #import "MSAlertController.h"
 #import "AppCenterDelegateObjC.h"
-#import "MSLogFilter.h"
+#import "MSEventFilter.h"
 
 @import AppCenter;
 @import AppCenterAnalytics;
@@ -144,21 +144,12 @@
   return [[[MSCrashes lastSessionCrashReport] device] carrierCountry];
 }
 
-#pragma mark - MSLogFilter section.
-- (BOOL) isLogFilterEnabled {
-  return [MSLogFilter isEnabled];
+#pragma mark - MSEventFilter section.
+- (BOOL) isEventFilterEnabled {
+  return [MSEventFilter isEnabled];
 }
-- (void) setLogFilterEnabled:(BOOL)isEnabled {
-  [MSLogFilter setEnabled:isEnabled];
-}
-- (void) filterLogType:(NSString*)logType {
-  [MSLogFilter filterLogType:logType];
-}
-- (void) unfilterLogType:(NSString*)logType {
-  [MSLogFilter unfilterLogType:logType];
-}
-- (BOOL) isFilteringLogType:(NSString*)logType {
-  return [MSLogFilter isFilteringLogType:logType];
+- (void) setEventFilterEnabled:(BOOL)isEnabled {
+  [MSEventFilter setEnabled:isEnabled];
 }
 
 @end
